+++ php/Terminus/Models/SshKey.php
@@ -5,7 +5,30 @@ namespace Terminus\Models;
 class SshKey extends TerminusModel {
 
   /**
-   * Returns the hex for this key
+   * Deletes a specific SSH key
+   *
+   * @return array
+   */
+  public function delete() {
+    $response = $this->request->request(
+      'users/' . $this->user->id . '/keys/' . $this->get('id'),
+      ['method' => 'delete',]
+    );
+  }
+
+  /**
+   * Returns the comment for this SSH key
+   *
+   * @return string
+   */
+  public function getComment() {
+    $key_parts = explode(' ', $this->get('key'));
+    $comment   = $key_parts[2];
+    return $comment;
+  }
+
+  /**
+   * Returns the hex for this SSH key
    *
    * @return string
    */